+++ packages/effect-pg-boss/src/index.ts
@@ -1,0 +1,33 @@
+export {
+	type DefineQueueOptions,
+	type DefineScheduleOptions,
+	defineQueue,
+	defineSchedule,
+	type JobContext,
+	type JobPayloadSchema,
+	type JobRegistration,
+	type QueueDefinition,
+	type QueueWorker,
+	type ScheduleDefinition,
+	type ScheduledWorker,
+} from "./definition.js";
+export {
+	PgBossError,
+	type PgBossOperation,
+	PgBossPayloadError,
+} from "./error.js";
+export {
+	deadLetterQueueName,
+	type JobsHealth,
+	type QueueHealth,
+} from "./health.js";
+export type {
+	PgBossClient,
+	PgBossClientFactory,
+} from "./internal/client.js";
+export {
+	makePgBoss,
+	type PgBossDefinition,
+	type PgBossLayerOptions,
+	type PgBossService,
+} from "./service.js";